Refactor: Refactor the structure
