+++ src/components/Routes/BusRoutes/RouteForm/RouteForm.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { Plus } from "lucide-react";
-import { MapLocationPicker } from "../../MapLocationPicker/MapLocationPicker";
+import { MapLocationPicker } from "../../../MapLocationPicker/MapLocationPicker";
 import {routeFormStyles } from "./routeForm.styles";
 
 interface RouteFormProps {
